Refactor SQL comment block classes to improve structure and indentation handling
